+++ src/components/inviteAuth/LoginComponent.js
@@ -1,0 +1,11 @@
+class LoginComponent {
+    get loginButtonInvite() { return $('button[data-testid=login-button]'); }
+
+    async clickLoginInvite() {
+        await this.loginButtonInvite.waitForDisplayed();
+        await this.loginButtonInvite.click();
+    }
+    
+}
+
+module.exports = new LoginComponent();
