+++ src/pages/AuthenticationPage.js
@@ -1,10 +1,10 @@
-const AuthenticationComponent = require('../components/AuthenticationComponent');
 const LoginFormComponent = require('../components/login/LoginFormComponent');
+const LoginComponent = require('../components/inviteAuth/LoginComponent');
 
 class AuthenticationPage {
     constructor() {
-        this.auth = AuthenticationComponent;
         this.loginForm = LoginFormComponent;
+        this.loginInvite = LoginComponent;
     }
 }
